Refactor: create assignment refactoring
