+++ .gitignore
@@ -2,3 +2,4 @@ dev
 node_modules
 .next
 .env
+.DS_Store
